+++ src/main/java/org/kumoricon/registration/model/attendee/CheckInByHourDTO.java
@@ -4,28 +4,53 @@ import java.time.OffsetDateTime;
 
 public class CheckInByHourDTO {
     private final OffsetDateTime start;
-    private final Integer preregCheckIn;
-    private final Integer atConCheckIn;
+    private final Integer attendeePreRegCheckedIn;
+    private final Integer attendeeAtConCheckedIn;
+    private final Integer vipPreRegCheckedIn;
+    private final Integer vipAtConCheckedIn;
+    private final Integer specialityPreRegCheckedIn;
+    private final Integer specialityAtConCheckedIn;
+    private final Integer staffCheckedIn;
 
-    public CheckInByHourDTO(OffsetDateTime start, Integer preregCheckIn, Integer atConCheckIn) {
+    public CheckInByHourDTO(OffsetDateTime start,
+                            Integer attendeePreRegCheckedIn, Integer attendeeAtConCheckedIn,
+                            Integer vipPreRegCheckedIn, Integer vipAtConCheckedIn,
+                            Integer specialtyPreRegCheckedIn, Integer specialityAtConCheckedIn,
+                            Integer staffCheckedIn) {
         this.start = start;
-        this.preregCheckIn = preregCheckIn;
-        this.atConCheckIn = atConCheckIn;
+        this.attendeePreRegCheckedIn = attendeePreRegCheckedIn;
+        this.attendeeAtConCheckedIn = attendeeAtConCheckedIn;
+        this.vipPreRegCheckedIn = vipPreRegCheckedIn;
+        this.vipAtConCheckedIn = vipAtConCheckedIn;
+        this.specialityPreRegCheckedIn = specialtyPreRegCheckedIn;
+        this.specialityAtConCheckedIn = specialityAtConCheckedIn;
+        this.staffCheckedIn = staffCheckedIn;
     }
 
     public OffsetDateTime getStart() {
         return start;
     }
 
-    public Integer getPreregCheckIn() {
-        return preregCheckIn;
+    public Integer getAttendeePreRegCheckedIn() {
+        return attendeePreRegCheckedIn;
     }
 
-    public Integer getAtConCheckIn() {
-        return atConCheckIn;
+    public Integer getAttendeeAtConCheckedIn() {
+        return attendeeAtConCheckedIn;
     }
+
+    public Integer getVipPreRegCheckedIn() { return vipPreRegCheckedIn; }
+
+    public Integer getVipAtConCheckedIn() { return vipAtConCheckedIn; }
+
+    public Integer getSpecialityPreRegCheckedIn() { return specialityPreRegCheckedIn; }
+
+    public Integer getSpecialityAtConCheckedIn() { return specialityAtConCheckedIn; }
+
+    public Integer getStaffCheckedIn() { return staffCheckedIn;}
 
     public Integer getTotal() {
-        return preregCheckIn + atConCheckIn;
+        return attendeePreRegCheckedIn + attendeeAtConCheckedIn + vipPreRegCheckedIn + vipAtConCheckedIn +
+                specialityPreRegCheckedIn + specialityAtConCheckedIn + staffCheckedIn;
     }
 }